ATR-824: renamed PX1005 analyzer and tests names to show that it supports action handlers
