+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/TypoInViewAndActionHandlerName/Sources/ViewDelegate/TypoInViewDelegateName_Bad_Expected.cs
