+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/TypoInViewAndActionHandlerName/Sources/ViewDelegate/TypoInViewDelegateName_Good_DifferentNames.cs
