+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/TypoInViewAndActionHandlerName/Sources/ViewDelegate/TypoInViewDelegateName_GraphExtension_Bad_Expected.cs
