+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/TypoInViewAndActionHandlerName/TypoInViewAndActionHandlerNameTests.cs
@@ -1,48 +1,52 @@
-﻿using System;
+﻿#nullable enable
+
+using System;
+
 using Acuminator.Analyzers.StaticAnalysis;
 using Acuminator.Analyzers.StaticAnalysis.PXGraph;
-using Acuminator.Analyzers.StaticAnalysis.TypoInViewDelegateName;
+using Acuminator.Analyzers.StaticAnalysis.TypoInViewAndActionHandlerName;
 using Acuminator.Tests.Helpers;
 using Acuminator.Tests.Verification;
 using Acuminator.Utilities;
+
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Xunit;
 
-namespace Acuminator.Tests.Tests.StaticAnalysis.TypoInViewDelegateName
+namespace Acuminator.Tests.Tests.StaticAnalysis.TypoInViewAndActionDelegateName
 {
-    public class TypoInViewDelegateNameTests : CodeFixVerifier
+    public class TypoInViewAndActionHandlerNameTests : CodeFixVerifier
 	{
 		protected override CodeFixProvider GetCSharpCodeFixProvider() => new TypoInViewDelegateNameFix();
 
 		protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer() =>
 			 new PXGraphAnalyzer(CodeAnalysisSettings.Default, 
-				 new TypoInViewDelegateNameAnalyzer());
+				 new TypoInViewAndActionHandlerNameAnalyzer());
 
 		[Theory]
-        [EmbeddedFileData("TypoInViewDelegateName_Good_SameName.cs")] 
+        [EmbeddedFileData(@"ViewDelegate\TypoInViewDelegateName_Good_SameName.cs")] 
 		public void TestDiagnostic_ShouldNotShowDiagnostic_SameName(string actual)
         {
             VerifyCSharpDiagnostic(actual);
         }
 
 	    [Theory]
-	    [EmbeddedFileData("TypoInViewDelegateName_Good_DifferentNames.cs")]
+	    [EmbeddedFileData(@"ViewDelegate\TypoInViewDelegateName_Good_DifferentNames.cs")]
 	    public void TestDiagnostic_ShouldNotShowDiagnostic_DifferentNames(string actual)
 	    {
 		    VerifyCSharpDiagnostic(actual);
 	    }
 
 	    [Theory]
-	    [EmbeddedFileData("TypoInViewDelegateName_Good_Override.cs")]
+	    [EmbeddedFileData(@"ViewDelegate\TypoInViewDelegateName_Good_Override.cs")]
 	    public void TestDiagnostic_ShouldNotShowDiagnostic_Override(string actual)
 	    {
 		    VerifyCSharpDiagnostic(actual);
 	    }
 
 		[Theory]
-        [EmbeddedFileData("TypoInViewDelegateName_Bad.cs")]
+        [EmbeddedFileData(@"ViewDelegate\TypoInViewDelegateName_Bad.cs")]
         public void TestDiagnostic(string actual)
         {
             VerifyCSharpDiagnostic(actual,
@@ -50,15 +54,15 @@ namespace Acuminator.Tests.Tests.StaticAnalysis.TypoInViewDelegateName
         }
 
 	    [Theory]
-	    [EmbeddedFileData("TypoInViewDelegateName_Bad.cs",
-						  "TypoInViewDelegateName_Bad_Expected.cs")]
+	    [EmbeddedFileData(@"ViewDelegate\TypoInViewDelegateName_Bad.cs",
+						  @"ViewDelegate\TypoInViewDelegateName_Bad_Expected.cs")]
 		public void TestCodeFix(string actual, string expected)
 	    {
 		    VerifyCSharpFix(actual, expected);
 	    }
 
 		[Theory]
-		[EmbeddedFileData("TypoInViewDelegateName_GraphExtension_Bad.cs")]
+		[EmbeddedFileData(@"ViewDelegate\TypoInViewDelegateName_GraphExtension_Bad.cs")]
 		public void TestDiagnostic_GraphExtension(string actual)
 		{
 			VerifyCSharpDiagnostic(actual, 
@@ -66,12 +70,12 @@ namespace Acuminator.Tests.Tests.StaticAnalysis.TypoInViewDelegateName
 		}
 
 		[Theory]
-		[EmbeddedFileData("TypoInViewDelegateName_GraphExtension_Bad_Expected.cs")]
+		[EmbeddedFileData(@"ViewDelegate\TypoInViewDelegateName_GraphExtension_Bad_Expected.cs")]
 		public void TestDiagnostic_GraphExtension_ShouldNotShowDiagnostic(string actual) => VerifyCSharpDiagnostic(actual);
 
 		[Theory]
-		[EmbeddedFileData("TypoInViewDelegateName_GraphExtension_Bad.cs",
-						  "TypoInViewDelegateName_GraphExtension_Bad_Expected.cs")]
+		[EmbeddedFileData(@"ViewDelegate\TypoInViewDelegateName_GraphExtension_Bad.cs",
+						  @"ViewDelegate\TypoInViewDelegateName_GraphExtension_Bad_Expected.cs")]
 		public void TestCodeFix_GraphExtension(string actual, string expected) => VerifyCSharpFix(actual, expected);
 	}
 }
